refactor: split main.rs into git/pr/render modules
Pure code move along the seams created by the prior two commits: git
(GitRepo + mmap cache), pr (cache codec, parse, fetch adapters), render
(StatusView, gather, components). main.rs keeps config, input structs,
shared utils, and main(). Cross-module items are pub(crate). Tests moved
alongside the code they cover. No behavior change; output byte-identical.

main.rs 2218 -> ~620 lines.
